refactor(bookmark): introduce BaseBookmark and specific bookmark entities for monument, mark, occurrence, and evidence
